Refactor compose files
