+++ app/src/main/java/ru/neexol/rtut/presentation/screens/notes/ui/NotesList.kt
@@ -1,4 +1,4 @@
-package ru.neexol.rtut.presentation.screens.notes
+package ru.neexol.rtut.presentation.screens.notes.ui
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.border
@@ -24,40 +24,7 @@ import ru.neexol.rtut.R
 import ru.neexol.rtut.data.notes.models.Note
 
 @Composable
-fun Notes(vm: NotesViewModel, navController: NavController) {
-	val uiState = vm.uiState
-	Box {
-		Column(Modifier.padding(horizontal = 20.dp)) {
-			LessonBar(vm.lesson, vm.isPublicType) {
-				vm.isPublicType = it
-			}
-			uiState.notes?.let { notes ->
-				NotesList(
-					if (vm.isPublicType) notes.second else notes.first,
-					vm.author,
-					navController
-				)
-			}
-			Spacer(Modifier.height(86.dp))
-		}
-		FloatingActionButton(
-			modifier = Modifier
-				.align(Alignment.BottomEnd)
-				.padding(20.dp),
-			shape = RoundedCornerShape(16.dp),
-			elevation = FloatingActionButtonDefaults.elevation(0.dp, 0.dp, 0.dp, 0.dp),
-			onClick = {
-//				vm.isPublicType = selectedType == NoteType.PUBLIC
-				navController.navigate("edit")
-			}
-		) {
-			Icon(painter = painterResource(R.drawable.ic_add_24), contentDescription = null)
-		}
-	}
-}
-
-@Composable
-private fun NotesList(notes: List<Note>, author: String, navController: NavController) {
+internal fun NotesList(notes: List<Note>, author: String, navController: NavController) {
 	if (notes.isEmpty()) {
 		val color = MaterialTheme.colors.onSurface.copy(alpha = ContentAlpha.medium)
 		Box(
