+++ app/src/main/java/ru/neexol/rtut/presentation/screens/schedule/ui/ScheduleScreen.kt
@@ -1,27 +1,20 @@
-package ru.neexol.rtut.presentation.screens.schedule
+package ru.neexol.rtut.presentation.screens.schedule.ui
 
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringArrayResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.pager.ExperimentalPagerApi
-import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.PagerState
 import com.google.accompanist.pager.rememberPagerState
 import kotlinx.coroutines.launch
-import ru.neexol.rtut.R
 import ru.neexol.rtut.data.lessons.models.Lesson
-import ru.neexol.rtut.data.lessons.models.LessonTime
-import ru.neexol.rtut.presentation.components.LessonItem
-import ru.neexol.rtut.presentation.components.PagerTopBar
+import ru.neexol.rtut.presentation.screens.schedule.ScheduleViewModel
 import java.math.BigDecimal
 
 @ExperimentalMaterialApi
@@ -90,53 +83,4 @@ private suspend fun syncScroll(pair: Pair<PagerState, PagerState>?) {
 				divideAndRemainder[1].toFloat(),
 			)
 		}
-}
-
-@ExperimentalPagerApi
-@Composable
-private fun WeekPagerBar(state: PagerState, items: List<String>, onTitleClick: () -> Unit) {
-	PagerTopBar(
-		state = state,
-		title = stringResource(R.string.week_letter),
-		items = items,
-		onTitleClick = onTitleClick
-	)
-}
-
-@ExperimentalPagerApi
-@Composable
-private fun DayPagerBar(state: PagerState, onTitleClick: () -> Unit) {
-	PagerTopBar(
-		state = state,
-		title = stringResource(R.string.day_letter),
-		items = stringArrayResource(R.array.days_cut).toList(),
-		onTitleClick = onTitleClick,
-		isLast = true
-	)
-}
-
-@ExperimentalMaterialApi
-@ExperimentalPagerApi
-@Composable
-private fun LessonsPager(
-	state: PagerState,
-	lessons: List<List<List<Lesson?>>>,
-	times: List<LessonTime>,
-	week: Int,
-	onLessonClick: (Lesson, String) -> Unit
-) {
-	HorizontalPager(
-		state = state,
-		count = 6
-	) { day ->
-		LazyColumn(
-			modifier = Modifier.fillMaxSize(),
-			contentPadding = PaddingValues(vertical = 14.dp, horizontal = 20.dp),
-			verticalArrangement = Arrangement.spacedBy(14.dp),
-		) {
-			itemsIndexed(lessons[week][day]) { number, lesson ->
-				LessonItem(lesson, times[number]) { onLessonClick(lesson!!, (week + 1).toString()) }
-			}
-		}
-	}
 }
