+++ app/src/main/java/ru/neexol/rtut/presentation/screens/settings/ui/SettingsScreen.kt
@@ -1,4 +1,4 @@
-package ru.neexol.rtut.presentation.screens.settings
+package ru.neexol.rtut.presentation.screens.settings.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -9,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import ru.neexol.rtut.presentation.screens.settings.SettingsViewModel
 
 @Composable
 fun SettingsScreen(vm: SettingsViewModel = hiltViewModel()) {
